Displaying an alert when a new notification is received and the app is open
